+++ lib/src/code_generators/swagger_requests_generator.dart
@@ -339,19 +339,23 @@ class SwaggerRequestsGenerator {
     required String modelPostfix,
     required SwaggerRoot root,
   }) {
+    final format = parameter.schema?.format ?? '';
+
     if (parameter.inParameter == kHeader) {
-      return _mapParameterName(kString, '');
+      return _mapParameterName(kString, format, '');
     } else if (parameter.items?.enumValues.isNotEmpty == true ||
         parameter.schema?.enumValues.isNotEmpty == true) {
       return _getEnumParameterTypeName(
           parameterName: parameter.name, path: path, requestType: requestType);
     } else if (parameter.items?.type.isNotEmpty == true) {
-      return _mapParameterName(parameter.items!.type, modelPostfix).asList();
+      return _mapParameterName(parameter.items!.type, format, modelPostfix)
+          .asList();
     } else if (parameter.items?.ref.isNotEmpty == true) {
       if (_isEnumRefParameter(parameter, root)) {
         return parameter.items!.ref.getRef().asEnum();
       }
-      return _mapParameterName(parameter.items!.ref.getRef(), modelPostfix)
+      return _mapParameterName(
+              parameter.items!.ref.getRef(), format, modelPostfix)
           .asList();
     } else if (parameter.schema?.items?.ref.isNotEmpty == true) {
       if (_isEnumRefParameter(parameter, root)) {
@@ -363,25 +367,33 @@ class SwaggerRequestsGenerator {
         return parameter.schema!.ref.getRef().asEnum();
       }
 
+      if (_isEnumRef(parameter.schema!.ref.getUnformattedRef(), root)) {
+        return parameter.schema!.ref.getRef().asEnum();
+      }
+
       if (parameter.schema!.items != null || parameter.schema!.type == kArray) {
         return (parameter.schema!.ref.getRef() + modelPostfix).asList();
       }
       return (parameter.schema!.ref.getRef() + modelPostfix);
     } else if (parameter.schema?.type == kArray &&
         parameter.schema?.items?.type.isNotEmpty == true) {
-      return _mapParameterName(parameter.schema!.items!.type, '').asList();
+      return _mapParameterName(parameter.schema!.items!.type, format, '')
+          .asList();
     } else if (parameter.schema?.anyOf.firstOrNull?.type.isNotEmpty == true) {
-      return _mapParameterName(parameter.schema!.anyOf.first.type, '');
+      return _mapParameterName(parameter.schema!.anyOf.first.type, format, '');
     }
 
-    final neededType = parameter.type.isNotEmpty
+    var neededType = parameter.type.isNotEmpty
         ? parameter.type
         : parameter.schema?.type ?? kObject.pascalCase;
 
-    return _mapParameterName(neededType, modelPostfix);
+    return _mapParameterName(neededType, format, modelPostfix);
   }
 
-  String _mapParameterName(String name, String modelPostfix) {
+  String _mapParameterName(String name, String format, String modelPostfix) {
+    if (name == kInteger && format == kInt64) {
+      return kNum;
+    }
     return kBasicTypesMap[name] ?? name.pascalCase + modelPostfix;
   }
 
@@ -515,7 +527,8 @@ class SwaggerRequestsGenerator {
 
     final neededSchema = schemas[neededSchemaKey]!;
 
-    if (neededSchema.type == kString && neededSchema.enumValues.isNotEmpty) {
+    if ((neededSchema.type == kString || neededSchema.type == kInteger) &&
+        neededSchema.enumValues.isNotEmpty) {
       return true;
     }
 
@@ -641,7 +654,7 @@ class SwaggerRequestsGenerator {
         return null;
       }
 
-      return _mapParameterName(mappedArrayType, modelPostfix).asList();
+      return _mapParameterName(mappedArrayType, '', modelPostfix).asList();
     }
 
     return kBasicTypesMap[responseType] ?? responseType + modelPostfix;